test: move cross-agent redirect tests into agentconn_test.go
